Don't change the username to the validated preferred username.
We used to do this because the username was changed as part of this call,
but it no longer is -- it's a just a idempotent duplicate check. Instead,
post the credentials as the user typed them and let `create_wp_user`
change the username -- something it is capable of doing in the event the
user has JS disabled.