First check the local WP user table before checking the API.
This means that, in order to tell if the user is new, we first
check the WP user table for someone with the same username. If
one exists, we assume it's them and let them authenticate. If
not, then we make a round trip to the API to see if they are
someone who hasn't yet joined.

This fixes an issue where local users -- really just user #1 --
could not log in if they didn't have the BeenHereBefore cookie
set. It probably also eliminates the need to set that cookie at
all, as all it does it save an API round-trip for users that
log in with a member number instead of username.